style: object storage 패키지 수정 스토리지 접근 관련 dto 분류
