+++ src/main/java/com/nhnacademy/booklay/server/dto/stroage/response/auth/Access.java
@@ -1,14 +1,16 @@
-package com.nhnacademy.booklay.server.dto.stroage.response;
+package com.nhnacademy.booklay.server.dto.stroage.response.auth;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @NoArgsConstructor
 @Getter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Tenant {
-
-    private String id;
+@ToString
+public class Access {
+    private Token token;
+    private User user;
 
 }
